% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjusted_rand}
\alias{adjusted_rand}
\title{Adjusted Rand index between two partitions}
\usage{
adjusted_rand(a, b)
}
\arguments{
\item{a, b}{integer vectors of group labels of equal length.}
}
\value{
the adjusted Rand index; 1 for identical partitions, 0 in
expectation for independent ones.
}
\description{
Used to score recovery of the true partition in the simulations
(paper Tables 2 and 3).
}
\examples{
adjusted_rand(c(1, 1, 2, 2), c(2, 2, 1, 1))  # same partition, relabelled
adjusted_rand(c(1, 1, 2, 2), c(1, 2, 1, 2))
}
